more pr feedback fixes
