+++ fixlib/fixlib/baseresources.py
@@ -12,10 +12,10 @@ from typing import Dict, Iterator, List, ClassVar, Optional, TypedDict, Any, Typ
 from collections import defaultdict
 
 from attr import resolve_types
-from attrs import define, field, Factory
+from attrs import define, field, Factory, frozen, evolve
 from prometheus_client import Counter, Summary
 
-from fixlib.json import from_json as _from_json, to_json as _to_json
+from fixlib.json import from_json as _from_json, to_json as _to_json, to_json_str
 from fixlib.logger import log
 from fixlib.types import Json
 from fixlib.utils import make_valid_timestamp, utc_str
@@ -1615,6 +1615,66 @@ class PolicySourceKind(StrEnum):
     Principal = "principal"  # e.g. IAM user, attached policy
     Group = "group"  # policy comes from an IAM group
     Resource = "resource"  # e.g. s3 bucket policy
+
+
+
+
+ResourceConstraint = str
+
+ConditionString = str
+
+
+@frozen
+class PolicySource:
+    kind: PolicySourceKind
+    uri: str
+
+
+class HasResourcePolicy(ABC):
+    # returns a list of all policies that affects the resource (inline, attached, etc.)
+    def resource_policy(self, builder: Any) -> List[Tuple[PolicySource, Json]]:
+        raise NotImplementedError
+
+
+@frozen
+class PermissionCondition:
+    # if nonempty and any evals to true, access is granted, otherwise implicitly denied
+    allow: Optional[Tuple[ConditionString, ...]] = None
+    # if nonempty and any is evals to false, access is implicitly denied
+    boundary: Optional[Tuple[ConditionString, ...]] = None
+    # if nonempty and any evals to true, access is explicitly denied
+    deny: Optional[Tuple[ConditionString, ...]] = None
+
+
+@frozen
+class PermissionScope:
+    source: PolicySource
+    constraints: Tuple[ResourceConstraint, ...]  # aka resource constraints
+    conditions: Optional[PermissionCondition] = None
+
+    def with_deny_conditions(self, deny_conditions: List[Json]) -> "PermissionScope":
+        c = self.conditions or PermissionCondition()
+        return evolve(self, conditions=evolve(c, deny=tuple([to_json_str(c) for c in deny_conditions])))
+
+    def with_boundary_conditions(self, boundary_conditions: List[Json]) -> "PermissionScope":
+        c = self.conditions or PermissionCondition()
+        return evolve(self, conditions=evolve(c, boundary=tuple([to_json_str(c) for c in boundary_conditions])))
+
+    def has_no_condititons(self) -> bool:
+        if self.conditions is None:
+            return True
+
+        if self.conditions.allow is None and self.conditions.boundary is None and self.conditions.deny is None:
+            return True
+
+        return False
+
+
+@frozen
+class AccessPermission:
+    action: str
+    level: str
+    scopes: Tuple[PermissionScope, ...]
 
 
 resolve_types(BaseResource)  # noqa
